rzutowanie rozmiaru akcji - wyciszenie warninga+backtrace
git-svn-id: svn+ssh://kadu.net/home/kadu/repos/kadu/trunk@4541 fea5205d-6df5-0310-bef7-ea67e47da9a5

+++ kadu-core/toolbar.cpp
@@ -429,7 +429,7 @@ void ToolBar::addDefaultAction(const QString &toolbar, const QString &actionName
 	action.showLabel = showLabel;
 	action.button = 0;
 
-	if (index >= actions.size())
+	if (index >= (int)actions.size())
 	{
 		kdebugm(KDEBUG_ERROR, "requested action index (%d) >= actions size (%u)!\n", index, actions.size());
 		printBacktrace("requested action index >= actions size!");
